Video refactor
