+++ source/input.h
@@ -1,6 +1,7 @@
 #ifndef INPUT_H
 #define INPUT_H
 
+void inputHandlerLoop(void* dummy);
 void handleInput();
 
 #endif
