+++ source/video.h
@@ -1,0 +1,6 @@
+#ifndef _VIDEO_H
+#define _VIDEO_H
+
+int handleVid(void);
+
+#endif
